feat: add "modified preferred" filter
- add filter to select attributions that were previously preferred but have been modified
- show auditing option chip when preferred attribution was modified with option to revert to original state
- show ungrouped autocomplete options with dense padding
