+++ src/shared/text.ts
@@ -67,6 +67,7 @@ export const text = {
     currentlyPreferred: 'Currently Preferred',
     excludedFromNotice: 'Excluded from Notice',
     followUp: 'Needs Follow-Up',
+    modifiedPreferred: 'Modified Previously Preferred',
     needsReview: 'Needs Review by QA',
     preselected: 'Pre-selected',
     previouslyPreferred: 'Previously Preferred',
@@ -90,6 +91,7 @@ export const text = {
     excludedFromNotice: 'Excluded from Notice',
     firstParty: 'First Party',
     lowConfidence: 'Low Confidence',
+    modifiedPreferred: 'Modified Previously Preferred',
     needsFollowUp: 'Needs Follow-Up',
     needsReview: 'Needs Review by QA',
     preSelected: 'Pre-selected',
